+++ test/src/Connector-test.js
@@ -1,4 +1,4 @@
-import Connector from '../../src/Connector';
+import Connector from '../../src/Connector.js';
 
 describe('Connector', () => {
   it('can be instantiated without new', () => {